dropme! squashed commits from feature/2102-2
2102: Cleanup: Remove constructors from game config types

We remove the constructors and use these types as simple structs. On top
of that we enable move semantics in more places.

2102: Cleanup: Use std::variant for texture package config

Previously we used an enum to differentiate the different cases. With
variant, we can save one alternative and make the code safer and less
awkward to use.

2102: Cleanup: Turn GameConfig into a struct

2102: Make member variables non-mutable

2102: Enable more move semantics

2102: Cleanup: Code style

2102: Store value data using shared_ptr

Since values are immutable, but are copied around a lot, this optimizes
performance a bit.

2102: Cleanup: code style

2102: Add comparison operators for expressions

2102: Make Expression immutable

We want to make Expression immutable so that we can store the expression
by shared pointer internally. This allows us to refer to expressions
from values later on.

The only non-const member function was optimize(), which now returns the
optimized expression instead of optimizing the expression in place.

Using shared_ptr to store the expression makes it possible to use all
defaulted copy / move constructors / assignments and destructor.

2102: Extract concrete expression types into separate unit

We need to extract these types into a separate unit in order to remove
the Value.h include from Expression.h. This would cause a circular
include in a following commit where we need to include Expression.h in
Value.h

2102: Remove converting constructor for array values

This constructor was used in only one place, and it's a bit misleading
since it doesn't set line / column to sensible values.

2102: Store originating expression in value

This change allows us to access the expression that "generated" a value.

2102: Remove Value::undefined() and Value::null() functions

We replace these functions with explicit comparisons to their singleton
values.

2102: Cleanup: add missing operator

2102: Cleanup: use free functions for comparison operators

2102: Add deref_iterator

This iterator is useful when working with collections of pointers. It
wraps another iterator and just dereferences that iterator twice when
it is itself dereferenced.

2102: Cleanup: add missing comparison operators for tag types

2102: Add stream insertion operators for Tag related classes

2102: Add stream insertion operators for game engine related classes

2102: Add stream insertion operators for compilation related classes

2102: Add stream insertion operator for brush face attribute classes

2102: Add stream insertion operators for game config

2102: Add support for configuring model scale in the game config

We add support for setting a scale expression in the game configuration.
This expression is evaluated at runtime against an entities' properties.
The scale expression configured in the game configuration is used only
if the model definition doesn't have its own scale expression.

2102: Cleanup: code style

2102: Cleanup: hide implementation details

2102: Cleanup: remove default arguments

2102: Add tests for ModelDefinition and remove redundant tests

The test cases we remove in the parser tests are redundant because they
test the functionality of the ModelDefinition class, not the parser
itself. We only leave one test each to ensure that parsing EL model
expressions works at all.

2102: Return Undefined for unknown entity properties

This is necessary to make scale expressions work as expected, i.e., if a
model definition has a scale expression that accesses an entity property
that an entity doesn't have, then the global scale should be returned.

2102: Add support for specifying scale in model expressions

2102: Cleanup: code style

2102: Cleanup: remove unused forward declaration

2102: Cleanup: rename PropertyKeys and PropertyNames

All other related types have the prefix "Entity", so we adapt these
names to be more consistent.

2102: Recompute cached entity properties eagerly

In the following commits, we will add the model transformation to the
cached entity properties. Once we start to consider the model scale in
the model transformation, we need to access the default model scale
expression from the game config, which we cannot access lazily.
Therefore, we change the updating to the cached entity properties to be
done eagerly here.

2102: Cache model transformation

2102: Cleanup: Remove delegating functions and access contained object

2102: Cleanup: Test structure

2102: Introduce EntityPropertyConfig

EntityPropertyConfig encapsulates any information that controls how
entity properties behave. Currently, it only contains a global scale
expression that is used if a model definition doesn't have a scale
expression itself, but in the future, it may also contain info about
the semantics of special entity property keys such as target or
targetname. This would then allow us to move away from hard coding
property keys.

EntityPropertyConfig must be passed to any member of Entity that
modifies entity properties in some way. Nodes can access a global config
object by calling Node::entityPropertyConfig, since the world node
stores that global config.

2102: Add model scale support to Quake 3 config

2102: Cleanup: Remove unused function

2102: Update manual for model scale